Split cst.rs parser into sub-modules
Splits the monolithic 3035-line parser into focused sub-modules:

- cst.rs (550 lines + tests): struct, new(), parse(), helpers, pattern
- cst/items.rs (803 lines): item, import, const, fn, type, for, trait
- cst/exprs.rs (961 lines): expressions, match, constructors, call args
- cst/types.rs (110 lines): type expressions
- cst/jsx.rs (157 lines): JSX elements, props, children

Uses Rust 2018+ module style (cst.rs + cst/ directory, no mod.rs).
Purely structural - no behavioral changes.
