+++ src/cst.rs
@@ -1,3 +1,8 @@
+mod exprs;
+mod items;
+mod jsx;
+mod types;
+
 use crate::lexer::span::Span;
 use crate::lexer::token::{Token, TokenKind};
 use crate::syntax::{SyntaxKind, SyntaxNode, token_kind_to_syntax};
@@ -66,2023 +71,6 @@ impl<'src> CstParser<'src> {
         Parse {
             green_node: self.builder.finish(),
             errors: self.errors,
-        }
-    }
-
-    // ── Items ────────────────────────────────────────────────────
-
-    fn parse_item(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-
-        // Handle export prefix
-        let exported = self.at(TokenKind::Export);
-        if exported {
-            self.bump(); // export
-            self.eat_trivia();
-        }
-
-        match self.current_kind() {
-            Some(TokenKind::Import) if !exported => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_import();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Import) => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ERROR.into());
-                self.error("cannot export an import statement");
-                self.bump();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Const) => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_const_decl();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Fn) if !self.peek_is(TokenKind::LeftParen) => {
-                // `fn name(...)` is a function declaration; `fn(...)` is a lambda expression
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_function_decl();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Async)
-                if self.peek_is(TokenKind::Fn)
-                    && self.peek_nth_non_trivia_kind(2) != Some(TokenKind::LeftParen) =>
-            {
-                // `async fn name(...)` is a function declaration; `async fn(...)` is a lambda
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_function_decl();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Opaque) | Some(TokenKind::Type) => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_type_decl();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::For) => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_for_block();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Trait) => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_trait_decl();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Use) if !exported => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_use_decl();
-                self.builder.finish_node();
-            }
-            _ if !exported && self.at_identifier("test") && self.peek_is_string() => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ITEM.into());
-                self.parse_test_block();
-                self.builder.finish_node();
-            }
-            _ if exported => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::ERROR.into());
-                self.error("expected declaration after 'export'");
-                self.builder.finish_node();
-            }
-            _ => {
-                self.builder
-                    .start_node_at(checkpoint, SyntaxKind::EXPR_ITEM.into());
-                self.parse_expr();
-                self.builder.finish_node();
-            }
-        }
-    }
-
-    // ── Import ────────────────────────────────────────────────────
-
-    fn parse_import(&mut self) {
-        self.builder.start_node(SyntaxKind::IMPORT_DECL.into());
-        self.expect(TokenKind::Import);
-        self.eat_trivia();
-
-        // `import trusted { ... }` — module-level trusted
-        if self.at_identifier("trusted") {
-            self.bump(); // trusted (emitted as IDENT token)
-            self.eat_trivia();
-        }
-
-        if self.at(TokenKind::LeftBrace) {
-            self.bump(); // {
-            self.eat_trivia();
-            self.parse_comma_separated(Self::parse_import_specifier_or_for, TokenKind::RightBrace);
-            self.expect(TokenKind::RightBrace);
-            self.eat_trivia();
-        }
-
-        // `from` is required with specifiers, optional for bare imports
-        if self.at(TokenKind::From) {
-            self.bump();
-            self.eat_trivia();
-        }
-        self.expect_kind(TokenKind::String("".into()));
-
-        self.builder.finish_node();
-    }
-
-    /// Parse either a regular import specifier or a `for Type` import specifier.
-    fn parse_import_specifier_or_for(&mut self) {
-        if self.at(TokenKind::For) {
-            // `for Type` import specifier
-            self.builder
-                .start_node(SyntaxKind::IMPORT_FOR_SPECIFIER.into());
-            self.bump(); // `for`
-            self.eat_trivia();
-            self.expect_ident(); // type name
-            self.builder.finish_node();
-        } else {
-            self.parse_import_specifier();
-        }
-    }
-
-    fn parse_import_specifier(&mut self) {
-        self.builder.start_node(SyntaxKind::IMPORT_SPECIFIER.into());
-        // `trusted foo` — per-specifier trusted
-        if self.at_identifier("trusted") && self.peek_is_ident() {
-            self.bump(); // trusted
-            self.eat_trivia();
-        }
-        self.expect_ident();
-        self.eat_trivia();
-
-        // Check for `as alias` — "as" is a banned keyword but used contextually here
-        if self.at_identifier("as")
-            || self.at(TokenKind::Banned(crate::lexer::token::BannedKeyword::As))
-        {
-            self.bump();
-            self.eat_trivia();
-            self.expect_ident();
-        }
-
-        self.builder.finish_node();
-    }
-
-    // ── Const Declaration ────────────────────────────────────────
-
-    fn parse_const_decl(&mut self) {
-        self.builder.start_node(SyntaxKind::CONST_DECL.into());
-        self.expect(TokenKind::Const);
-        self.eat_trivia();
-
-        if self.at(TokenKind::LeftBracket) {
-            // Array destructuring
-            self.bump();
-            self.eat_trivia();
-            self.parse_comma_separated(Self::expect_ident_item, TokenKind::RightBracket);
-            self.expect(TokenKind::RightBracket);
-        } else if self.at(TokenKind::LeftBrace) {
-            // Object destructuring: { a, b } or { a: x, b: y }
-            self.bump();
-            self.eat_trivia();
-            self.parse_comma_separated(Self::parse_destructure_field, TokenKind::RightBrace);
-            self.expect(TokenKind::RightBrace);
-        } else if self.at(TokenKind::LeftParen) && self.is_const_tuple_destructuring() {
-            // Tuple destructuring: const (a, b) = ...
-            self.bump();
-            self.eat_trivia();
-            self.parse_comma_separated(Self::expect_ident_item, TokenKind::RightParen);
-            self.expect(TokenKind::RightParen);
-        } else {
-            self.expect_ident();
-        }
-        self.eat_trivia();
-
-        // Optional type annotation
-        if self.at(TokenKind::Colon) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_type_expr();
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::Equal);
-        self.eat_trivia();
-        self.parse_expr();
-
-        self.builder.finish_node();
-    }
-
-    // ── Use Declaration ─────────────────────────────────────────
-
-    fn parse_use_decl(&mut self) {
-        self.builder.start_node(SyntaxKind::USE_DECL.into());
-        self.expect(TokenKind::Use);
-        self.eat_trivia();
-
-        // Optional binding: `use x <- expr` or `use (a, b) <- expr` or `use <- expr`
-        if !self.at(TokenKind::LeftArrow) {
-            if self.at(TokenKind::LeftParen) {
-                // Tuple destructuring: `use (a, b) <- expr`
-                self.bump();
-                self.eat_trivia();
-                self.parse_comma_separated(Self::expect_ident_item, TokenKind::RightParen);
-                self.expect(TokenKind::RightParen);
-            } else {
-                self.expect_ident();
-            }
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::LeftArrow);
-        self.eat_trivia();
-        self.parse_expr();
-
-        self.builder.finish_node();
-    }
-
-    // ── Function Declaration ────────────────────────────────────
-
-    fn parse_function_decl(&mut self) {
-        self.builder.start_node(SyntaxKind::FUNCTION_DECL.into());
-
-        if self.at(TokenKind::Async) {
-            self.bump();
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::Fn);
-        self.eat_trivia();
-        self.expect_ident();
-        self.eat_trivia();
-
-        // Optional type parameters: <T, U>
-        if self.at(TokenKind::LessThan) {
-            self.bump(); // <
-            self.eat_trivia();
-            self.parse_comma_separated(Self::expect_ident_item, TokenKind::GreaterThan);
-            self.expect(TokenKind::GreaterThan);
-            self.eat_trivia();
-        }
-
-        if self.at(TokenKind::Equal) {
-            // `fn name = expr` — derived function binding (pointfree style)
-            self.bump(); // =
-            self.eat_trivia();
-            self.parse_expr();
-        } else {
-            // `fn name(params) { body }` — standard function declaration
-            self.expect(TokenKind::LeftParen);
-            self.eat_trivia();
-            self.parse_comma_separated(Self::parse_param, TokenKind::RightParen);
-            self.expect(TokenKind::RightParen);
-            self.eat_trivia();
-
-            // Optional return type
-            if self.at(TokenKind::ThinArrow) {
-                self.bump();
-                self.eat_trivia();
-                self.parse_type_expr();
-                self.eat_trivia();
-            }
-
-            self.parse_block_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_param(&mut self) {
-        self.builder.start_node(SyntaxKind::PARAM.into());
-
-        if self.at(TokenKind::LeftBrace) {
-            // Destructured param: { name, age } or { name: n, age: a }
-            self.bump(); // {
-            self.eat_trivia();
-            self.parse_comma_separated(Self::parse_destructure_field, TokenKind::RightBrace);
-            self.expect(TokenKind::RightBrace);
-            self.eat_trivia();
-        } else if self.at(TokenKind::LeftParen) {
-            // Tuple destructured param: (a, b)
-            self.bump(); // (
-            self.eat_trivia();
-            self.parse_comma_separated(Self::expect_ident_item, TokenKind::RightParen);
-            self.expect(TokenKind::RightParen);
-            self.eat_trivia();
-        } else if self.at(TokenKind::SelfKw) {
-            self.bump(); // self
-            self.eat_trivia();
-        } else if self.at(TokenKind::Underscore) {
-            self.bump(); // _
-            self.eat_trivia();
-        } else {
-            self.expect_ident();
-            self.eat_trivia();
-        }
-
-        if self.at(TokenKind::Colon) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_type_expr();
-            self.eat_trivia();
-        }
-
-        if self.at(TokenKind::Equal) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    // ── Type Declaration ────────────────────────────────────────
-
-    fn parse_type_decl(&mut self) {
-        self.builder.start_node(SyntaxKind::TYPE_DECL.into());
-
-        if self.at(TokenKind::Opaque) {
-            self.bump();
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::Type);
-        self.eat_trivia();
-        self.expect_ident();
-        self.eat_trivia();
-
-        // Optional type parameters: <T, U>
-        if self.at(TokenKind::LessThan) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_comma_separated(Self::expect_ident_item, TokenKind::GreaterThan);
-            self.expect(TokenKind::GreaterThan);
-            self.eat_trivia();
-        }
-
-        // New syntax: `type Name { ... }` for records/unions/newtypes
-        // Old syntax: `type Name = ...` for aliases and string literal unions
-        if self.at(TokenKind::LeftBrace) {
-            self.parse_type_body_in_braces();
-        } else {
-            self.expect(TokenKind::Equal);
-            self.eat_trivia();
-            self.parse_type_def_after_eq();
-        }
-
-        // Optional deriving clause: `deriving (Display)`
-        self.eat_trivia();
-        if self.at(TokenKind::Deriving) {
-            self.builder.start_node(SyntaxKind::DERIVING_CLAUSE.into());
-            self.bump(); // consume `deriving`
-            self.eat_trivia();
-            self.expect(TokenKind::LeftParen);
-            self.eat_trivia();
-            self.parse_comma_separated(Self::expect_ident_item, TokenKind::RightParen);
-            self.expect(TokenKind::RightParen);
-            self.builder.finish_node();
-        }
-
-        self.builder.finish_node();
-    }
-
-    /// Parse type body inside `{ }`: disambiguate between record, union, and newtype.
-    fn parse_type_body_in_braces(&mut self) {
-        // Peek at first non-trivia token inside `{` to disambiguate:
-        // - `|` → union variants
-        // - lowercase ident + `:` → record fields
-        // - `...` → record fields (spread)
-        // - `}` → empty record
-        // - anything else → newtype wrapper
-        let first_inside = self.peek_inside_brace();
-
-        match first_inside {
-            Some(TokenKind::VerticalBar) => {
-                self.builder.start_node(SyntaxKind::TYPE_DEF_UNION.into());
-                self.bump(); // {
-                self.eat_trivia();
-                self.parse_union_variants_inner();
-                self.expect(TokenKind::RightBrace);
-                self.builder.finish_node();
-            }
-            Some(TokenKind::DotDotDot) => {
-                // Record with spread
-                self.builder.start_node(SyntaxKind::TYPE_DEF_RECORD.into());
-                self.parse_record_fields();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Identifier(name)) if name.starts_with(char::is_lowercase) => {
-                // Peek further: if followed by `:`, it's a record field.
-                // Otherwise it's a newtype (e.g. `type OrderId { number }`)
-                if self.peek_inside_brace_second() == Some(TokenKind::Colon) {
-                    self.builder.start_node(SyntaxKind::TYPE_DEF_RECORD.into());
-                    self.parse_record_fields();
-                    self.builder.finish_node();
-                } else {
-                    // Newtype wrapping a lowercase type like `number`, `string`, `boolean`
-                    self.builder.start_node(SyntaxKind::TYPE_DEF_UNION.into());
-                    self.bump(); // {
-                    self.eat_trivia();
-                    self.builder.start_node(SyntaxKind::VARIANT_FIELD.into());
-                    self.parse_type_expr();
-                    self.builder.finish_node();
-                    self.eat_trivia();
-                    self.expect(TokenKind::RightBrace);
-                    self.builder.finish_node();
-                }
-            }
-            Some(TokenKind::RightBrace) => {
-                // Empty record: `type Foo {}`
-                self.builder.start_node(SyntaxKind::TYPE_DEF_RECORD.into());
-                self.parse_record_fields();
-                self.builder.finish_node();
-            }
-            _ => {
-                // Newtype: `type OrderId { number }`
-                // Parse as single-variant union matching the type name
-                self.builder.start_node(SyntaxKind::TYPE_DEF_UNION.into());
-                self.bump(); // {
-                self.eat_trivia();
-                // Synthesize a variant with the type's name — the lowerer
-                // will pick up the inner type expression as a variant field
-                self.builder.start_node(SyntaxKind::VARIANT_FIELD.into());
-                self.parse_type_expr();
-                self.builder.finish_node();
-                self.eat_trivia();
-                self.expect(TokenKind::RightBrace);
-                self.builder.finish_node();
-            }
-        }
-    }
-
-    /// Peek at the first non-trivia token after the current `{`.
-    fn peek_inside_brace(&self) -> Option<TokenKind> {
-        let mut i = self.pos + 1;
-        while i < self.tokens.len() {
-            if !self.tokens[i].kind.is_trivia() {
-                return Some(self.tokens[i].kind.clone());
-            }
-            i += 1;
-        }
-        None
-    }
-
-    /// Peek at the second non-trivia token after the current `{`.
-    fn peek_inside_brace_second(&self) -> Option<TokenKind> {
-        let mut i = self.pos + 1;
-        let mut count = 0;
-        while i < self.tokens.len() {
-            if !self.tokens[i].kind.is_trivia() {
-                count += 1;
-                if count == 2 {
-                    return Some(self.tokens[i].kind.clone());
-                }
-            }
-            i += 1;
-        }
-        None
-    }
-
-    /// Parse after `=`: aliases and string literal unions only.
-    fn parse_type_def_after_eq(&mut self) {
-        if self.at_string_literal_union() {
-            self.parse_string_literal_union();
-        } else {
-            self.builder.start_node(SyntaxKind::TYPE_DEF_ALIAS.into());
-            self.parse_type_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    /// Parse union variants inside `{ }`. The `{` is already consumed, `}` is consumed by caller.
-    fn parse_union_variants_inner(&mut self) {
-        while self.at_pipe_in_union() {
-            self.builder.start_node(SyntaxKind::VARIANT.into());
-            self.bump(); // |
-            self.eat_trivia();
-            self.expect_ident();
-            self.eat_trivia();
-
-            // Variant fields now use { } instead of ( )
-            if self.at(TokenKind::LeftBrace) {
-                self.bump(); // {
-                self.eat_trivia();
-                self.parse_comma_separated(Self::parse_variant_field, TokenKind::RightBrace);
-                self.expect(TokenKind::RightBrace);
-                self.eat_trivia();
-            }
-
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_string_literal_union(&mut self) {
-        self.builder
-            .start_node(SyntaxKind::TYPE_DEF_STRING_UNION.into());
-
-        // First string literal
-        self.bump(); // string
-        self.eat_trivia();
-
-        // Parse remaining `| "string"` pairs
-        while self.at(TokenKind::VerticalBar) {
-            self.bump(); // |
-            self.eat_trivia();
-            if self.at(TokenKind::String("".into())) {
-                self.bump(); // string
-                self.eat_trivia();
-            } else {
-                self.error("expected string literal after `|` in string literal union");
-                break;
-            }
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_variant_field(&mut self) {
-        self.builder.start_node(SyntaxKind::VARIANT_FIELD.into());
-
-        // Check if this is a named field: `name: Type`
-        if self.is_ident() && self.peek_is(TokenKind::Colon) {
-            self.bump(); // name
-            self.eat_trivia();
-            self.bump(); // :
-            self.eat_trivia();
-            self.parse_type_expr();
-        } else {
-            self.parse_type_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_record_fields(&mut self) {
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-        self.parse_comma_separated(Self::parse_record_entry, TokenKind::RightBrace);
-        self.expect(TokenKind::RightBrace);
-    }
-
-    fn parse_record_entry(&mut self) {
-        // Check for spread: `...TypeName` or `...Generic<T>`
-        if self.at(TokenKind::DotDotDot) {
-            self.builder.start_node(SyntaxKind::RECORD_SPREAD.into());
-            self.bump(); // consume `...`
-            self.eat_trivia();
-            self.parse_type_expr();
-            self.builder.finish_node();
-            return;
-        }
-
-        self.parse_record_field();
-    }
-
-    fn parse_record_field(&mut self) {
-        self.builder.start_node(SyntaxKind::RECORD_FIELD.into());
-        self.expect_ident();
-        self.eat_trivia();
-        self.expect(TokenKind::Colon);
-        self.eat_trivia();
-        self.parse_type_expr();
-        self.eat_trivia();
-
-        if self.at(TokenKind::Equal) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    // ── For Blocks ──────────────────────────────────────────────
-
-    /// Parse a for-block: `for Type { fn ... }`.
-    fn parse_for_block(&mut self) {
-        self.builder.start_node(SyntaxKind::FOR_BLOCK.into());
-
-        self.expect(TokenKind::For);
-        self.eat_trivia();
-
-        // Parse the type name (e.g., `User`, `Array<T>`)
-        self.parse_type_expr();
-        self.eat_trivia();
-
-        // Optional trait bound: `for User: Display { ... }`
-        if self.at(TokenKind::Colon) {
-            self.bump(); // :
-            self.eat_trivia();
-            self.expect_ident(); // trait name
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-
-        // Parse function declarations inside the block (with optional export)
-        while !self.at(TokenKind::RightBrace) && !self.at_end() {
-            if self.at(TokenKind::Export) {
-                self.bump();
-                self.eat_trivia();
-            }
-            if self.at(TokenKind::Fn) || self.at(TokenKind::Async) {
-                self.parse_for_block_function();
-                self.eat_trivia();
-            } else {
-                self.error("expected `fn` inside for block");
-                self.bump();
-                self.eat_trivia();
-            }
-        }
-
-        self.expect(TokenKind::RightBrace);
-
-        self.builder.finish_node();
-    }
-
-    // ── Trait Declarations ────────────────────────────────────────
-
-    fn parse_trait_decl(&mut self) {
-        self.builder.start_node(SyntaxKind::TRAIT_DECL.into());
-
-        self.expect(TokenKind::Trait);
-        self.eat_trivia();
-
-        self.expect_ident(); // trait name
-        self.eat_trivia();
-
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-
-        // Parse method declarations inside the trait
-        while !self.at(TokenKind::RightBrace) && !self.at_end() {
-            if self.at(TokenKind::Fn) {
-                self.parse_trait_method();
-                self.eat_trivia();
-            } else {
-                self.error("expected `fn` inside trait");
-                self.bump();
-                self.eat_trivia();
-            }
-        }
-
-        self.expect(TokenKind::RightBrace);
-
-        self.builder.finish_node();
-    }
-
-    fn parse_trait_method(&mut self) {
-        self.builder.start_node(SyntaxKind::FUNCTION_DECL.into());
-
-        self.expect(TokenKind::Fn);
-        self.eat_trivia();
-        self.expect_ident();
-        self.eat_trivia();
-
-        self.expect(TokenKind::LeftParen);
-        self.eat_trivia();
-        self.parse_comma_separated(Self::parse_for_block_param, TokenKind::RightParen);
-        self.expect(TokenKind::RightParen);
-        self.eat_trivia();
-
-        // Optional return type
-        if self.at(TokenKind::ThinArrow) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_type_expr();
-            self.eat_trivia();
-        }
-
-        // Optional body (default implementation)
-        if self.at(TokenKind::LeftBrace) {
-            self.parse_block_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_for_block_function(&mut self) {
-        self.builder.start_node(SyntaxKind::FUNCTION_DECL.into());
-
-        if self.at(TokenKind::Async) {
-            self.bump();
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::Fn);
-        self.eat_trivia();
-        self.expect_ident();
-        self.eat_trivia();
-
-        self.expect(TokenKind::LeftParen);
-        self.eat_trivia();
-        self.parse_comma_separated(Self::parse_for_block_param, TokenKind::RightParen);
-        self.expect(TokenKind::RightParen);
-        self.eat_trivia();
-
-        // Optional return type
-        if self.at(TokenKind::ThinArrow) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_type_expr();
-            self.eat_trivia();
-        }
-
-        self.parse_block_expr();
-
-        self.builder.finish_node();
-    }
-
-    fn parse_for_block_param(&mut self) {
-        self.builder.start_node(SyntaxKind::PARAM.into());
-
-        if self.at(TokenKind::SelfKw) {
-            // `self` parameter — bump as an ident-like token
-            self.bump();
-        } else {
-            self.expect_ident();
-            self.eat_trivia();
-
-            if self.at(TokenKind::Colon) {
-                self.bump();
-                self.eat_trivia();
-                self.parse_type_expr();
-                self.eat_trivia();
-            }
-
-            if self.at(TokenKind::Equal) {
-                self.bump();
-                self.eat_trivia();
-                self.parse_expr();
-            }
-        }
-
-        self.builder.finish_node();
-    }
-
-    // ── Test Blocks ──────────────────────────────────────────────
-
-    fn parse_test_block(&mut self) {
-        self.builder.start_node(SyntaxKind::TEST_BLOCK.into());
-
-        // `test` is a contextual keyword (an identifier)
-        self.bump(); // consume "test" identifier
-        self.eat_trivia();
-
-        // Test name (string literal)
-        self.expect_kind(TokenKind::String("".into()));
-        self.eat_trivia();
-
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-
-        // Parse test body: assert statements and expressions
-        while !self.at(TokenKind::RightBrace) && !self.at_end() {
-            let prev_pos = self.pos;
-            if self.at(TokenKind::Assert) {
-                self.parse_assert_stmt();
-            } else {
-                self.parse_expr();
-            }
-            self.eat_trivia();
-            if self.pos == prev_pos && !self.at_end() {
-                self.bump();
-            }
-        }
-
-        self.expect(TokenKind::RightBrace);
-
-        self.builder.finish_node();
-    }
-
-    fn parse_assert_stmt(&mut self) {
-        self.builder.start_node(SyntaxKind::ASSERT_EXPR.into());
-
-        self.expect(TokenKind::Assert);
-        self.eat_trivia();
-        self.parse_expr();
-
-        self.builder.finish_node();
-    }
-
-    // ── Type Expressions ────────────────────────────────────────
-
-    fn parse_type_expr(&mut self) {
-        self.builder.start_node(SyntaxKind::TYPE_EXPR.into());
-
-        // Function type: (params) -> ReturnType or () -> ReturnType
-        if self.at(TokenKind::LeftParen) && self.is_paren_function_type() {
-            self.parse_function_type();
-        }
-        // Unit type: ()
-        else if self.at(TokenKind::LeftParen) && self.peek_is(TokenKind::RightParen) {
-            self.bump(); // (
-            self.eat_trivia();
-            self.bump(); // )
-        }
-        // Function type: fn(params) -> ReturnType (old syntax — error)
-        else if self.at(TokenKind::Fn) && self.peek_is(TokenKind::LeftParen) {
-            self.builder.start_node(SyntaxKind::ERROR.into());
-            self.error("function types use arrow syntax: `(T) -> U` instead of `fn(T) -> U`");
-            self.bump(); // fn
-            self.builder.finish_node();
-        }
-        // Tuple type: (T, U) — paren with comma, no `->` after `)`
-        else if self.at(TokenKind::LeftParen) && self.is_paren_tuple_type() {
-            self.bump(); // (
-            self.eat_trivia();
-            self.parse_comma_separated(Self::parse_type_expr, TokenKind::RightParen);
-            self.expect(TokenKind::RightParen);
-        }
-        // Tuple: [T, U]
-        else if self.at(TokenKind::LeftBracket) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_comma_separated(Self::parse_type_expr, TokenKind::RightBracket);
-            self.expect(TokenKind::RightBracket);
-        }
-        // Record type: { ... }
-        else if self.at(TokenKind::LeftBrace) {
-            self.parse_record_fields();
-        }
-        // String literal type (e.g. ComponentProps<"div">)
-        else if self.at(TokenKind::String("".into())) {
-            self.bump();
-        }
-        // typeof <ident> or typeof Module.value
-        else if self.at(TokenKind::Typeof) {
-            self.bump();
-            self.eat_trivia();
-            self.expect_ident();
-            self.eat_trivia();
-            while self.at(TokenKind::Dot) {
-                self.bump();
-                self.eat_trivia();
-                self.expect_ident();
-                self.eat_trivia();
-            }
-        }
-        // Named type
-        else {
-            self.expect_ident();
-            self.eat_trivia();
-
-            // Dotted names (e.g. JSX.Element)
-            while self.at(TokenKind::Dot) {
-                self.bump();
-                self.eat_trivia();
-                self.expect_ident();
-                self.eat_trivia();
-            }
-
-            // Type arguments: <T, U>
-            if self.at(TokenKind::LessThan) {
-                self.bump();
-                self.eat_trivia();
-                self.parse_comma_separated(Self::parse_type_expr, TokenKind::GreaterThan);
-                self.expect(TokenKind::GreaterThan);
-                self.eat_trivia();
-            }
-        }
-
-        // Intersection: A & B & C — parse as flat list within this TYPE_EXPR node
-        while self.at(TokenKind::Amp) {
-            self.bump(); // &
-            self.eat_trivia();
-            self.parse_type_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_function_type(&mut self) {
-        self.expect(TokenKind::LeftParen);
-        self.eat_trivia();
-        self.parse_comma_separated(Self::parse_type_expr, TokenKind::RightParen);
-        self.expect(TokenKind::RightParen);
-        self.eat_trivia();
-        // Accept -> for function types; give helpful error if user writes =>
-        if self.at(TokenKind::FatArrow) {
-            self.error("function types use `->` not `=>`: write `(T) -> U`");
-            self.bump(); // consume => anyway to recover
-        } else {
-            self.expect(TokenKind::ThinArrow);
-        }
-        self.eat_trivia();
-        self.parse_type_expr();
-    }
-
-    // ── Expressions ─────────────────────────────────────────────
-
-    fn parse_expr(&mut self) {
-        self.parse_or_expr();
-    }
-
-    fn parse_or_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_and_expr();
-
-        while self.at(TokenKind::PipePipe) {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::BINARY_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_and_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_and_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_equality_expr();
-
-        while self.at(TokenKind::AmpAmp) {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::BINARY_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_equality_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_equality_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_pipe_expr();
-
-        while self.at(TokenKind::EqualEqual) || self.at(TokenKind::BangEqual) {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::BINARY_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_pipe_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_pipe_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_comparison_expr();
-
-        while self.at(TokenKind::Pipe) {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::PIPE_EXPR.into());
-            self.bump(); // |>
-            self.eat_trivia();
-
-            // Pipe into match: `x |> match { ... }`
-            if self.at(TokenKind::Match) {
-                self.parse_subjectless_match_expr();
-            } else {
-                self.parse_comparison_expr();
-            }
-
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_comparison_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_additive_expr();
-
-        while (self.at(TokenKind::LessThan)
-            || self.at(TokenKind::GreaterThan)
-            || self.at(TokenKind::LessEqual)
-            || self.at(TokenKind::GreaterEqual))
-            && !self.preceded_by_newline()
-        {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::BINARY_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_additive_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_additive_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_multiplicative_expr();
-
-        while self.at(TokenKind::Plus) || self.at(TokenKind::Minus) {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::BINARY_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_multiplicative_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_multiplicative_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_unary_expr();
-
-        while self.at(TokenKind::Star) || self.at(TokenKind::Slash) || self.at(TokenKind::Percent) {
-            self.builder
-                .start_node_at(checkpoint, SyntaxKind::BINARY_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_unary_expr();
-            self.builder.finish_node();
-        }
-    }
-
-    fn parse_unary_expr(&mut self) {
-        match self.current_kind() {
-            Some(TokenKind::Bang) | Some(TokenKind::Minus) => {
-                self.builder.start_node(SyntaxKind::UNARY_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.parse_unary_expr();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::Await) => {
-                self.builder.start_node(SyntaxKind::AWAIT_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.parse_unary_expr();
-                self.builder.finish_node();
-            }
-            _ => self.parse_postfix_expr(),
-        }
-    }
-
-    fn parse_postfix_expr(&mut self) {
-        let checkpoint = self.builder.checkpoint();
-        self.parse_primary_expr();
-
-        loop {
-            self.eat_trivia();
-            match self.current_kind() {
-                Some(TokenKind::Question) => {
-                    self.builder
-                        .start_node_at(checkpoint, SyntaxKind::UNWRAP_EXPR.into());
-                    self.bump();
-                    self.builder.finish_node();
-                }
-                Some(TokenKind::Dot) => {
-                    self.builder
-                        .start_node_at(checkpoint, SyntaxKind::MEMBER_EXPR.into());
-                    self.bump();
-                    self.eat_trivia();
-                    // Accept identifiers, numbers (tuple .0, .1), banned keywords, and other keywords after `.`
-                    // (e.g., `Array.any(...)`, `Number.parse(...)`, `pair.0`)
-                    if self.is_ident()
-                        || matches!(
-                            self.current_kind(),
-                            Some(TokenKind::Number(_))
-                                | Some(TokenKind::Banned(_))
-                                | Some(TokenKind::Parse)
-                                | Some(TokenKind::Match)
-                                | Some(TokenKind::For)
-                                | Some(TokenKind::From)
-                                | Some(TokenKind::Type)
-                                | Some(TokenKind::Export)
-                                | Some(TokenKind::Import)
-                                | Some(TokenKind::Const)
-                                | Some(TokenKind::Fn)
-                                | Some(TokenKind::Async)
-                                | Some(TokenKind::Await)
-                                | Some(TokenKind::Trait)
-                                | Some(TokenKind::Collect)
-                                | Some(TokenKind::Deriving)
-                                | Some(TokenKind::When)
-                                | Some(TokenKind::SelfKw)
-                        )
-                    {
-                        self.bump();
-                    } else {
-                        self.expect_ident();
-                    }
-                    self.builder.finish_node();
-                }
-                Some(TokenKind::LeftBracket) => {
-                    self.builder
-                        .start_node_at(checkpoint, SyntaxKind::INDEX_EXPR.into());
-                    self.bump();
-                    self.eat_trivia();
-                    self.parse_expr();
-                    self.eat_trivia();
-                    self.expect(TokenKind::RightBracket);
-                    self.builder.finish_node();
-                }
-                Some(TokenKind::LessThan) if self.is_generic_call() => {
-                    // Generic call: `f<T>(args)` or `f<T, U>(args)`
-                    self.builder
-                        .start_node_at(checkpoint, SyntaxKind::CALL_EXPR.into());
-                    self.bump(); // <
-                    self.eat_trivia();
-                    self.parse_comma_separated(Self::parse_type_expr, TokenKind::GreaterThan);
-                    self.expect(TokenKind::GreaterThan);
-                    self.eat_trivia();
-                    self.expect(TokenKind::LeftParen);
-                    self.eat_trivia();
-                    self.parse_comma_separated(Self::parse_call_arg, TokenKind::RightParen);
-                    self.expect(TokenKind::RightParen);
-                    self.builder.finish_node();
-                }
-                Some(TokenKind::LeftParen) => {
-                    // Don't treat `(` on a new line as a call — it's a new expression
-                    if self.preceded_by_newline() {
-                        break;
-                    }
-                    // Check if it's a constructor (uppercase ident) — don't parse as call
-                    if self.is_uppercase_ident_at_checkpoint() {
-                        break;
-                    }
-                    self.builder
-                        .start_node_at(checkpoint, SyntaxKind::CALL_EXPR.into());
-                    self.bump(); // (
-                    self.eat_trivia();
-                    self.parse_comma_separated(Self::parse_call_arg, TokenKind::RightParen);
-                    self.expect(TokenKind::RightParen);
-                    self.builder.finish_node();
-                }
-                _ => break,
-            }
-        }
-    }
-
-    fn parse_primary_expr(&mut self) {
-        match self.current_kind() {
-            Some(TokenKind::Number(_)) => self.bump(),
-            Some(TokenKind::String(_)) => self.bump(),
-            Some(TokenKind::TemplateLiteral(_)) => self.bump(),
-            Some(TokenKind::Bool(_)) => self.bump(),
-            Some(TokenKind::Underscore) => self.bump(),
-            Some(TokenKind::None) => self.bump(),
-            Some(TokenKind::Clear) => self.bump(),
-            Some(TokenKind::Unchanged) => self.bump(),
-            Some(TokenKind::Todo) => self.bump(),
-            Some(TokenKind::Unreachable) => self.bump(),
-
-            Some(TokenKind::Ok) => {
-                self.builder.start_node(SyntaxKind::OK_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.expect(TokenKind::LeftParen);
-                self.eat_trivia();
-                self.parse_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::RightParen);
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Err) => {
-                self.builder.start_node(SyntaxKind::ERR_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.expect(TokenKind::LeftParen);
-                self.eat_trivia();
-                self.parse_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::RightParen);
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Some) => {
-                self.builder.start_node(SyntaxKind::SOME_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.expect(TokenKind::LeftParen);
-                self.eat_trivia();
-                self.parse_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::RightParen);
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Value) => {
-                self.builder.start_node(SyntaxKind::VALUE_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.expect(TokenKind::LeftParen);
-                self.eat_trivia();
-                self.parse_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::RightParen);
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Parse) => {
-                self.builder.start_node(SyntaxKind::PARSE_EXPR.into());
-                self.bump(); // parse
-                self.eat_trivia();
-                // parse<T> — type argument
-                self.expect(TokenKind::LessThan);
-                self.eat_trivia();
-                self.parse_type_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::GreaterThan);
-                self.eat_trivia();
-                // Optional (value) — may be absent in pipe context
-                if self.current_kind() == Some(TokenKind::LeftParen) {
-                    self.bump();
-                    self.eat_trivia();
-                    self.parse_expr();
-                    self.eat_trivia();
-                    self.expect(TokenKind::RightParen);
-                }
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Mock) => {
-                self.builder.start_node(SyntaxKind::MOCK_EXPR.into());
-                self.bump(); // mock
-                self.eat_trivia();
-                // mock<T> — type argument
-                self.expect(TokenKind::LessThan);
-                self.eat_trivia();
-                self.parse_type_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::GreaterThan);
-                self.eat_trivia();
-                // Optional (overrides) — named args for field overrides
-                if self.current_kind() == Some(TokenKind::LeftParen) {
-                    self.bump();
-                    self.eat_trivia();
-                    while self.current_kind() != Some(TokenKind::RightParen)
-                        && self.current_kind().is_some()
-                    {
-                        self.parse_call_arg();
-                        self.eat_trivia();
-                        if self.current_kind() == Some(TokenKind::Comma) {
-                            self.bump();
-                            self.eat_trivia();
-                        }
-                    }
-                    self.expect(TokenKind::RightParen);
-                }
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Try) => {
-                self.builder.start_node(SyntaxKind::TRY_EXPR.into());
-                self.bump(); // try
-                self.eat_trivia();
-                self.parse_expr();
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::Match) => self.parse_match_expr(),
-            Some(TokenKind::Collect) => {
-                self.builder.start_node(SyntaxKind::COLLECT_EXPR.into());
-                self.bump(); // collect
-                self.eat_trivia();
-                self.parse_block_expr();
-                self.builder.finish_node();
-            }
-            Some(TokenKind::LeftBrace) => {
-                if self.is_object_literal() {
-                    self.parse_object_literal();
-                } else {
-                    self.parse_block_expr();
-                }
-            }
-
-            Some(TokenKind::LeftBracket) => {
-                self.builder.start_node(SyntaxKind::ARRAY_EXPR.into());
-                self.bump();
-                self.eat_trivia();
-                self.parse_comma_separated(Self::parse_expr, TokenKind::RightBracket);
-                self.expect(TokenKind::RightBracket);
-                self.builder.finish_node();
-            }
-
-            Some(TokenKind::LeftParen) => {
-                if self.is_arrow_expr() {
-                    // Arrow closure: `(params) => body`
-                    self.parse_arrow_closure();
-                } else if self.peek_is(TokenKind::RightParen) {
-                    // Unit value: ()
-                    self.builder.start_node(SyntaxKind::TUPLE_EXPR.into());
-                    self.bump(); // (
-                    self.eat_trivia();
-                    self.bump(); // )
-                    self.builder.finish_node();
-                } else if self.is_paren_tuple_expr() {
-                    // Tuple: (expr, expr, ...)
-                    self.builder.start_node(SyntaxKind::TUPLE_EXPR.into());
-                    self.bump(); // (
-                    self.eat_trivia();
-                    self.parse_comma_separated(Self::parse_expr, TokenKind::RightParen);
-                    self.expect(TokenKind::RightParen);
-                    self.builder.finish_node();
-                } else {
-                    self.builder.start_node(SyntaxKind::GROUPED_EXPR.into());
-                    self.bump(); // (
-                    self.eat_trivia();
-                    self.parse_expr();
-                    self.eat_trivia();
-                    self.expect(TokenKind::RightParen);
-                    self.builder.finish_node();
-                }
-            }
-
-            Some(TokenKind::LessThan) => self.parse_jsx_element(),
-
-            Some(TokenKind::Dot) => {
-                self.parse_dot_shorthand();
-            }
-
-            Some(TokenKind::Async)
-                if self.peek_is(TokenKind::LeftParen) && self.is_async_arrow_expr() =>
-            {
-                // `async (params) => body`
-                self.parse_arrow_closure();
-            }
-
-            Some(TokenKind::Fn) if self.peek_is(TokenKind::LeftParen) => {
-                // `fn(params) expr` is the old syntax — emit error pointing to =>
-                self.builder.start_node(SyntaxKind::ERROR.into());
-                self.error("anonymous functions use arrow syntax: `(params) => body` instead of `fn(params) body`");
-                self.bump(); // fn
-                self.builder.finish_node();
-            }
-
-            // `self` keyword — treat as identifier in expression context
-            Some(TokenKind::SelfKw) => {
-                self.bump();
-            }
-
-            Some(TokenKind::Identifier(name)) => {
-                let name = name.clone();
-
-                // Uppercase + ( → constructor
-                if name.starts_with(char::is_uppercase) && self.peek_is(TokenKind::LeftParen) {
-                    self.parse_construct_expr();
-                    return;
-                }
-
-                // Qualified variant: `Filter.All` or `Route.Profile(id: "123")`
-                if name.starts_with(char::is_uppercase)
-                    && self.peek_is(TokenKind::Dot)
-                    && let Some(TokenKind::Identifier(variant_name)) =
-                        self.peek_nth_non_trivia_kind(2)
-                    && variant_name.starts_with(char::is_uppercase)
-                {
-                    // Check if there's a `(` after the variant name (3rd non-trivia)
-                    let has_args =
-                        matches!(self.peek_nth_non_trivia_kind(3), Some(TokenKind::LeftParen));
-
-                    // Emit as CONSTRUCT_EXPR for both unit and parameterized variants
-                    self.builder.start_node(SyntaxKind::CONSTRUCT_EXPR.into());
-                    self.bump(); // type name (Filter/Route)
-                    self.eat_trivia();
-                    self.bump(); // .
-                    self.eat_trivia();
-                    self.bump(); // variant name
-                    self.eat_trivia();
-
-                    if has_args {
-                        self.expect(TokenKind::LeftParen);
-                        self.eat_trivia();
-
-                        // Check for spread
-                        if self.at(TokenKind::DotDot) {
-                            self.builder.start_node(SyntaxKind::SPREAD_EXPR.into());
-                            self.bump();
-                            self.eat_trivia();
-                            self.parse_expr();
-                            self.builder.finish_node();
-                            self.eat_trivia();
-                            if self.at(TokenKind::Comma) {
-                                self.bump();
-                                self.eat_trivia();
-                            }
-                        }
-
-                        if !self.at(TokenKind::RightParen) {
-                            self.parse_comma_separated(Self::parse_call_arg, TokenKind::RightParen);
-                        }
-
-                        self.expect(TokenKind::RightParen);
-                    }
-
-                    self.builder.finish_node();
-                    return;
-                }
-
-                self.bump();
-            }
-
-            Some(TokenKind::Banned(_)) => {
-                self.builder.start_node(SyntaxKind::ERROR.into());
-                let kind = self.current_kind().unwrap();
-                if let TokenKind::Banned(banned) = kind {
-                    self.error(&format!(
-                        "banned keyword '{}': {}",
-                        banned.as_str(),
-                        banned.help_message()
-                    ));
-                }
-                self.bump();
-                self.builder.finish_node();
-            }
-
-            _ => {
-                self.builder.start_node(SyntaxKind::ERROR.into());
-                if let Some(kind) = self.current_kind() {
-                    self.error(&format!("unexpected token: {:?}", kind));
-                    self.bump();
-                }
-                self.builder.finish_node();
-            }
-        }
-    }
-
-    // ── Constructors ─────────────────────────────────────────────
-
-    fn parse_construct_expr(&mut self) {
-        self.builder.start_node(SyntaxKind::CONSTRUCT_EXPR.into());
-        self.bump(); // TypeName
-        self.eat_trivia();
-        self.expect(TokenKind::LeftParen);
-        self.eat_trivia();
-
-        // Check for spread: `..expr`
-        if self.at(TokenKind::DotDot) {
-            self.builder.start_node(SyntaxKind::SPREAD_EXPR.into());
-            self.bump();
-            self.eat_trivia();
-            self.parse_expr();
-            self.builder.finish_node();
-            self.eat_trivia();
-            if self.at(TokenKind::Comma) {
-                self.bump();
-                self.eat_trivia();
-            }
-        }
-
-        if !self.at(TokenKind::RightParen) {
-            self.parse_comma_separated(Self::parse_call_arg, TokenKind::RightParen);
-        }
-
-        self.expect(TokenKind::RightParen);
-        self.builder.finish_node();
-    }
-
-    // ── Call Arguments ───────────────────────────────────────────
-
-    fn parse_call_arg(&mut self) {
-        self.builder.start_node(SyntaxKind::ARG.into());
-
-        // Named arg: `label: expr` or punned `label:`
-        if self.is_ident() && self.peek_is(TokenKind::Colon) {
-            self.bump(); // label
-            self.eat_trivia();
-            self.bump(); // :
-
-            // Punning: `label:` without a value — next non-trivia is `)` or `,`
-            let next = self.next_non_trivia_kind();
-            let is_pun = matches!(
-                next,
-                Some(TokenKind::RightParen) | Some(TokenKind::Comma) | None
-            );
-            if !is_pun {
-                self.eat_trivia();
-                self.parse_expr();
-            }
-        } else {
-            self.parse_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    // ── Dot Shorthand ────────────────────────────────────────────
-
-    /// Parse `.field` or `.field op expr` dot shorthand expression.
-    fn parse_dot_shorthand(&mut self) {
-        self.builder.start_node(SyntaxKind::DOT_SHORTHAND.into());
-        self.expect(TokenKind::Dot);
-        self.eat_trivia();
-        self.expect_ident();
-        self.eat_trivia();
-
-        // Check for optional binary operator predicate
-        if self.at(TokenKind::EqualEqual)
-            || self.at(TokenKind::BangEqual)
-            || self.at(TokenKind::LessThan)
-            || self.at(TokenKind::GreaterThan)
-            || self.at(TokenKind::LessEqual)
-            || self.at(TokenKind::GreaterEqual)
-            || self.at(TokenKind::AmpAmp)
-            || self.at(TokenKind::PipePipe)
-            || self.at(TokenKind::Plus)
-            || self.at(TokenKind::Minus)
-            || self.at(TokenKind::Star)
-            || self.at(TokenKind::Slash)
-            || self.at(TokenKind::Percent)
-        {
-            self.bump(); // operator
-            self.eat_trivia();
-            self.parse_primary_expr();
-        }
-
-        self.builder.finish_node();
-    }
-
-    // ── Fn Lambda ────────────────────────────────────────────────
-
-    /// Parse `(params) => body` or `async (params) => body` arrow closure.
-    fn parse_arrow_closure(&mut self) {
-        self.builder.start_node(SyntaxKind::ARROW_EXPR.into());
-
-        // Optional `async` keyword
-        if self.at(TokenKind::Async) {
-            self.bump(); // async
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::LeftParen);
-        self.eat_trivia();
-        self.parse_comma_separated(Self::parse_param, TokenKind::RightParen);
-        self.expect(TokenKind::RightParen);
-        self.eat_trivia();
-        self.expect(TokenKind::FatArrow);
-        self.eat_trivia();
-        self.parse_expr();
-
-        self.builder.finish_node();
-    }
-
-    // ── Match Expression ─────────────────────────────────────────
-
-    fn parse_match_expr(&mut self) {
-        self.builder.start_node(SyntaxKind::MATCH_EXPR.into());
-        self.expect(TokenKind::Match);
-        self.eat_trivia();
-        self.parse_expr();
-        self.eat_trivia();
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-
-        while !self.at(TokenKind::RightBrace) && !self.at_end() {
-            let prev_pos = self.pos;
-            self.parse_match_arm();
-            self.eat_trivia();
-            if self.at(TokenKind::Comma) {
-                self.bump();
-                self.eat_trivia();
-            }
-            if self.pos == prev_pos && !self.at_end() {
-                self.bump();
-            }
-        }
-
-        self.expect(TokenKind::RightBrace);
-        self.builder.finish_node();
-    }
-
-    /// Parse `match { arms }` without a subject — used for `x |> match { ... }`.
-    fn parse_subjectless_match_expr(&mut self) {
-        self.builder.start_node(SyntaxKind::MATCH_EXPR.into());
-        self.expect(TokenKind::Match);
-        self.eat_trivia();
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-
-        while !self.at(TokenKind::RightBrace) && !self.at_end() {
-            let prev_pos = self.pos;
-            self.parse_match_arm();
-            self.eat_trivia();
-            if self.at(TokenKind::Comma) {
-                self.bump();
-                self.eat_trivia();
-            }
-            if self.pos == prev_pos && !self.at_end() {
-                self.bump();
-            }
-        }
-
-        self.expect(TokenKind::RightBrace);
-        self.builder.finish_node();
-    }
-
-    fn parse_match_arm(&mut self) {
-        self.builder.start_node(SyntaxKind::MATCH_ARM.into());
-        self.parse_pattern();
-        self.eat_trivia();
-
-        // Optional guard: `when expr`
-        if self.at(TokenKind::When) {
-            self.builder.start_node(SyntaxKind::MATCH_GUARD.into());
-            self.bump(); // consume `when`
-            self.eat_trivia();
-            self.parse_expr();
-            self.builder.finish_node();
-            self.eat_trivia();
-        }
-
-        self.expect(TokenKind::ThinArrow);
-        self.eat_trivia();
-        self.parse_expr();
-        self.builder.finish_node();
-    }
-
-    fn parse_pattern(&mut self) {
-        self.builder.start_node(SyntaxKind::PATTERN.into());
-
-        match self.current_kind() {
-            Some(TokenKind::Underscore) => {
-                self.bump();
-            }
-            Some(TokenKind::Bool(_)) => {
-                self.bump();
-            }
-            Some(TokenKind::String(_)) => {
-                self.bump();
-            }
-            Some(TokenKind::Minus) => {
-                // Negative number pattern: `-1`, `-3.14`
-                self.bump(); // -
-                self.eat_trivia();
-                if matches!(self.current_kind(), Some(TokenKind::Number(_))) {
-                    self.bump();
-                } else {
-                    self.error("expected number after '-' in pattern");
-                }
-            }
-            Some(TokenKind::Number(_)) => {
-                self.bump();
-                self.eat_trivia();
-                if self.at(TokenKind::DotDot) {
-                    self.bump();
-                    self.eat_trivia();
-                    // Expect number after ..
-                    if matches!(self.current_kind(), Some(TokenKind::Number(_))) {
-                        self.bump();
-                    } else {
-                        self.error("expected number after '..' in range pattern");
-                    }
-                }
-            }
-            Some(TokenKind::LeftBrace) => {
-                self.bump(); // {
-                self.eat_trivia();
-                self.parse_comma_separated(Self::parse_record_pattern_field, TokenKind::RightBrace);
-                self.expect(TokenKind::RightBrace);
-            }
-            Some(TokenKind::LeftBracket) => {
-                // Array pattern: [], [a, b], [first, ..rest]
-                self.bump(); // [
-                self.eat_trivia();
-                // Parse elements and optional rest pattern
-                while !self.at(TokenKind::RightBracket) && !self.at_end() {
-                    // Check for rest pattern: ..name
-                    if self.at(TokenKind::DotDot) {
-                        self.bump(); // ..
-                        self.eat_trivia();
-                        // Expect identifier or _ after ..
-                        if matches!(
-                            self.current_kind(),
-                            Some(TokenKind::Identifier(_)) | Some(TokenKind::Underscore)
-                        ) {
-                            self.bump();
-                        } else {
-                            self.error("expected identifier after '..' in array pattern");
-                        }
-                        self.eat_trivia();
-                        if self.at(TokenKind::Comma) {
-                            self.bump();
-                            self.eat_trivia();
-                        }
-                        break;
-                    }
-                    self.parse_pattern();
-                    self.eat_trivia();
-                    if self.at(TokenKind::Comma) {
-                        self.bump();
-                        self.eat_trivia();
-                    } else {
-                        break;
-                    }
-                }
-                self.expect(TokenKind::RightBracket);
-            }
-            Some(TokenKind::LeftParen) => {
-                // Tuple pattern: (x, y)
-                self.bump(); // (
-                self.eat_trivia();
-                self.parse_comma_separated(Self::parse_pattern, TokenKind::RightParen);
-                self.expect(TokenKind::RightParen);
-            }
-            Some(TokenKind::None) => {
-                self.bump();
-            }
-            Some(TokenKind::Ok) | Some(TokenKind::Err) | Some(TokenKind::Some) => {
-                self.bump();
-                self.eat_trivia();
-                if self.at(TokenKind::LeftParen) {
-                    self.bump();
-                    self.eat_trivia();
-                    self.parse_comma_separated(Self::parse_pattern, TokenKind::RightParen);
-                    self.expect(TokenKind::RightParen);
-                }
-            }
-            // Implicit variant pattern: .Variant or .Variant(args)
-            Some(TokenKind::Dot) if matches!(self.peek_nth_non_trivia_kind(1), Some(TokenKind::Identifier(n)) if n.starts_with(char::is_uppercase)) =>
-            {
-                self.bump(); // .
-                self.eat_trivia();
-                self.bump(); // Variant name
-                self.eat_trivia();
-                if self.at(TokenKind::LeftParen) {
-                    self.bump();
-                    self.eat_trivia();
-                    self.parse_comma_separated(Self::parse_pattern, TokenKind::RightParen);
-                    self.expect(TokenKind::RightParen);
-                }
-            }
-            Some(TokenKind::Identifier(name)) => {
-                let name = name.clone();
-                if name.starts_with(char::is_uppercase) {
-                    self.bump();
-                    self.eat_trivia();
-                    // Qualified variant pattern: `Type.Variant` or `Type.Variant(...)`
-                    if self.at(TokenKind::Dot) {
-                        self.bump(); // .
-                        self.eat_trivia();
-                        // Accept identifiers and keywords (Ok, Err, Some) after dot
-                        if self.is_ident()
-                            || matches!(
-                                self.current_kind(),
-                                Some(TokenKind::Ok)
-                                    | Some(TokenKind::Err)
-                                    | Some(TokenKind::Some)
-                                    | Some(TokenKind::None)
-                            )
-                        {
-                            self.bump();
-                        } else {
-                            self.expect_ident();
-                        }
-                        self.eat_trivia();
-                    }
-                    if self.at(TokenKind::LeftParen) {
-                        self.bump();
-                        self.eat_trivia();
-                        self.parse_comma_separated(Self::parse_pattern, TokenKind::RightParen);
-                        self.expect(TokenKind::RightParen);
-                    }
-                } else {
-                    self.bump();
-                }
-            }
-            _ => {
-                self.error(&format!(
-                    "unexpected token in pattern: {:?}",
-                    self.current_kind()
-                ));
-                if !self.at_end() {
-                    self.bump();
-                }
-            }
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_record_pattern_field(&mut self) {
-        self.expect_ident();
-        self.eat_trivia();
-        if self.at(TokenKind::Colon) {
-            self.bump();
-            self.eat_trivia();
-            self.parse_pattern();
-        }
-    }
-
-    // ── Object Literal ───────────────────────────────────────────
-
-    /// Check if the current `{` starts an object literal rather than a block.
-    /// An object literal has the form `{ ident: expr, ... }` or `{ ident, ... }` (shorthand).
-    fn is_object_literal(&self) -> bool {
-        // Look ahead past trivia after `{`
-        let mut i = self.pos + 1;
-        while i < self.tokens.len() && self.tokens[i].kind.is_trivia() {
-            i += 1;
-        }
-        if i >= self.tokens.len() {
-            return false;
-        }
-        // Must be an identifier (not a keyword)
-        if !matches!(self.tokens[i].kind, TokenKind::Identifier(_)) {
-            return false;
-        }
-        // Next non-trivia token after the ident must be `:` (key: value) or `,` or `}` (shorthand)
-        i += 1;
-        while i < self.tokens.len() && self.tokens[i].kind.is_trivia() {
-            i += 1;
-        }
-        if i >= self.tokens.len() {
-            return false;
-        }
-        matches!(
-            self.tokens[i].kind,
-            TokenKind::Colon | TokenKind::Comma | TokenKind::RightBrace
-        )
-    }
-
-    fn parse_object_literal(&mut self) {
-        self.builder.start_node(SyntaxKind::OBJECT_EXPR.into());
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-        self.parse_comma_separated(Self::parse_object_field, TokenKind::RightBrace);
-        self.expect(TokenKind::RightBrace);
-        self.builder.finish_node();
-    }
-
-    fn parse_object_field(&mut self) {
-        self.builder.start_node(SyntaxKind::OBJECT_FIELD.into());
-        self.expect_ident();
-        self.eat_trivia();
-        if self.at(TokenKind::Colon) {
-            self.bump(); // :
-            self.eat_trivia();
-            self.parse_expr();
-        }
-        // If no colon, it's shorthand: { name } means { name: name }
-        self.builder.finish_node();
-    }
-
-    // ── Block Expression ─────────────────────────────────────────
-
-    fn parse_block_expr(&mut self) {
-        self.builder.start_node(SyntaxKind::BLOCK_EXPR.into());
-        self.expect(TokenKind::LeftBrace);
-        self.eat_trivia();
-
-        while !self.at(TokenKind::RightBrace) && !self.at_end() {
-            let prev_pos = self.pos;
-            self.parse_item();
-            self.eat_trivia();
-            if self.pos == prev_pos && !self.at_end() {
-                self.bump();
-            }
-        }
-
-        self.expect(TokenKind::RightBrace);
-        self.builder.finish_node();
-    }
-
-    // ── JSX ──────────────────────────────────────────────────────
-
-    fn parse_jsx_element(&mut self) {
-        self.builder.start_node(SyntaxKind::JSX_ELEMENT.into());
-        self.expect(TokenKind::LessThan);
-        self.eat_trivia();
-
-        // Fragment: <>
-        if self.at(TokenKind::GreaterThan) {
-            self.bump(); // >
-            self.parse_jsx_children();
-            // Expect </>
-            self.expect(TokenKind::LessThan);
-            self.expect(TokenKind::Slash);
-            self.expect(TokenKind::GreaterThan);
-            self.builder.finish_node();
-            return;
-        }
-
-        self.expect_ident(); // tag name
-        self.eat_trivia();
-
-        // Props
-        while !self.at(TokenKind::GreaterThan) && !self.at(TokenKind::Slash) && !self.at_end() {
-            let prev_pos = self.pos;
-            self.parse_jsx_prop();
-            self.eat_trivia();
-            if self.pos == prev_pos && !self.at_end() {
-                // Safety: skip stuck token to prevent infinite loop.
-                self.error(&format!(
-                    "unexpected token in JSX element: {:?}",
-                    self.current_kind()
-                ));
-                self.bump();
-            }
-        }
-
-        // Self-closing: />
-        if self.at(TokenKind::Slash) {
-            self.bump();
-            self.expect(TokenKind::GreaterThan);
-            self.builder.finish_node();
-            return;
-        }
-
-        self.expect(TokenKind::GreaterThan);
-        self.parse_jsx_children();
-
-        // Closing tag: </Tag>
-        self.expect(TokenKind::LessThan);
-        self.expect(TokenKind::Slash);
-        self.eat_trivia();
-        self.expect_ident();
-        self.expect(TokenKind::GreaterThan);
-
-        self.builder.finish_node();
-    }
-
-    fn parse_jsx_prop(&mut self) {
-        // JSX spread: {...expr}
-        if self.at(TokenKind::LeftBrace) && self.peek_is(TokenKind::DotDotDot) {
-            self.builder.start_node(SyntaxKind::JSX_SPREAD_PROP.into());
-            self.bump(); // {
-            self.eat_trivia();
-            self.bump(); // ...
-            self.eat_trivia();
-            self.parse_expr();
-            self.eat_trivia();
-            self.expect(TokenKind::RightBrace);
-            self.builder.finish_node();
-            return;
-        }
-
-        self.builder.start_node(SyntaxKind::JSX_PROP.into());
-        // Accept identifiers and keywords as JSX prop names (e.g., type="text", for="id")
-        if self.is_ident() || self.is_keyword() {
-            self.bump();
-        } else {
-            self.expect_ident();
-        }
-        self.eat_trivia();
-
-        if self.at(TokenKind::Equal) {
-            self.bump();
-            self.eat_trivia();
-            if self.at(TokenKind::LeftBrace) {
-                self.bump();
-                self.eat_trivia();
-                self.parse_expr();
-                self.eat_trivia();
-                self.expect(TokenKind::RightBrace);
-            } else if matches!(self.current_kind(), Some(TokenKind::String(_))) {
-                self.bump();
-            } else {
-                self.error("expected '{' or string after '=' in JSX prop");
-            }
-        }
-
-        self.builder.finish_node();
-    }
-
-    fn parse_jsx_children(&mut self) {
-        loop {
-            // Check for closing tag
-            if self.at(TokenKind::LessThan) && self.peek_is(TokenKind::Slash) {
-                break;
-            }
-            if self.at_end() {
-                break;
-            }
-
-            let prev_pos = self.pos;
-            match self.current_kind() {
-                Some(TokenKind::LeftBrace) => {
-                    self.builder.start_node(SyntaxKind::JSX_EXPR_CHILD.into());
-                    self.bump();
-                    self.eat_trivia();
-                    // {/* comment */} — block comment already eaten as trivia,
-                    // so if the next token is `}` there's no expr to parse.
-                    if !self.at(TokenKind::RightBrace) {
-                        self.parse_expr();
-                        self.eat_trivia();
-                    }
-                    self.expect(TokenKind::RightBrace);
-                    self.builder.finish_node();
-                }
-                Some(TokenKind::LessThan) => {
-                    self.parse_jsx_element();
-                }
-                _ => {
-                    if self.is_jsx_text_token() {
-                        self.builder.start_node(SyntaxKind::JSX_TEXT.into());
-                        self.bump();
-                        while !self.at_end()
-                            && !self.at(TokenKind::LeftBrace)
-                            && !self.at(TokenKind::LessThan)
-                            && self.is_jsx_text_token()
-                        {
-                            // Eat whitespace between text tokens too
-                            self.bump();
-                        }
-                        self.builder.finish_node();
-                    } else {
-                        break;
-                    }
-                }
-            }
-            // Safety: if no progress was made, skip the stuck token.
-            if self.pos == prev_pos && !self.at_end() {
-                self.bump();
-            }
         }
     }
